app/requirements.txt
groq
llama-index
llama-index-readers-github
python-dotenv